2-dimensional and multi-variable options added
FossilOrigin-Name: af7da50ba1d23d43fef22fc80157ce1a8fac38ce25da5d62be1d489233a0fccf

+++ so08_qg_func.py
@@ -46,36 +46,17 @@ class SO08_qg_func():
 
     # get truth and make observation
     def get_true_and_obs(self):
-        #f = os.path.join(os.path.abspath(os.path.dirname(__file__)), \
-        #    "data/data.csv")
-        #truth = pd.read_csv(f)
-        #xt = truth.values.reshape(self.namax,self.nx)
-        truefile = self.truedir+f"/q{self.t0true:06d}.npy"
-        qt = np.zeros((self.na*self.nt+1,self.nx))
-        if not os.path.isfile(truefile):
-            logger.error(f"not exist {truefile}")
-            exit
-        else:
-            logger.info("read truth q")
-            qt[0,] = np.ravel(np.load(truefile))
-        truefile = self.truedir+f"/p{self.t0true:06d}.npy"
-        psit = np.zeros((self.na*self.nt+1,self.nx))
-        if not os.path.isfile(truefile):
-            logger.error(f"not exist {truefile}")
-            exit
-        else:
-            logger.info("read truth p")
-            psit[0,] = np.ravel(np.load(truefile))
-        #logger.debug("xt={}".format(xt))
         #observation settings
+        nvar = 1 #psi
         obs_offset = 10
         obs_interval = 27
-        obsloc = np.zeros((self.nobs,2))
+        obsloc = np.zeros((self.nobs,3)) #ivar,iloc,jloc
         iloc = obs_offset
         jloc = 2
         for i in range(self.nobs):
-            obsloc[i,0] = iloc
-            obsloc[i,1] = jloc
+            obsloc[i,0] = nvar
+            obsloc[i,1] = iloc
+            obsloc[i,2] = jloc
             iloc += obs_interval
             if iloc > self.ni-2:
                 iloc = iloc - self.ni + 2
@@ -84,23 +65,61 @@ class SO08_qg_func():
         obs_s = self.obs.get_sig()
         oberr = int(obs_s)
         obsfile="obs_{}_{}.npy".format(self.op, oberr)
+        yobs = np.zeros((self.na,self.nobs,4)) # var,location[i,j] and value
         
-        logger.info("start nature run and create obs")
-        logger.info("create obs")
-        yobs = np.zeros((self.na,self.nobs,3)) # location[i,j] and value
-        i=0
-        for icyc in range(self.na):
-            for it in range(self.nt):
-                q = qt[i,].reshape(self.ni,self.nj)
-                psi = psit[i,].reshape(self.ni,self.nj)
-                q = self.step(q,psi)
-                qt[i+1,] = np.ravel(q)
-                psit[i+1,] = np.ravel(psi)
-                i+=1
-            yobs[icyc,:,:2] = obsloc
-            yobs[icyc,:,2] = self.obs.add_noise(self.obs.h_operator(obsloc,psit[i,]))
-        np.save(obsfile, yobs)
+        fq = self.truedir+"/qtrue.npy"
+        fp = self.truedir+"/ptrue.npy"
+        if not os.path.isfile(fq) or not os.path.isfile(fp):
+            logger.info("create truth")
+            truefile = self.truedir+f"/q{self.t0true:06d}.npy"
+            qt = np.zeros((self.na*self.nt+1,self.nx))
+            if not os.path.isfile(truefile):
+                logger.error(f"not exist {truefile}")
+                exit()
+            else:
+                logger.info("read initial q")
+                qt[0,] = np.ravel(np.load(truefile))
+            truefile = self.truedir+f"/p{self.t0true:06d}.npy"
+            psit = np.zeros((self.na*self.nt+1,self.nx))
+            if not os.path.isfile(truefile):
+                logger.error(f"not exist {truefile}")
+                exit()
+            else:
+                logger.info("read initial p")
+                psit[0,] = np.ravel(np.load(truefile))
+        #logger.debug("xt={}".format(xt))
         
+            logger.info("start nature run and create obs")
+            i=0
+            for icyc in range(self.na):
+                for it in range(self.nt):
+                    q = qt[i,].reshape(self.ni,self.nj)
+                    psi = psit[i,].reshape(self.ni,self.nj)
+                    q = self.step(q,psi)
+                    qt[i+1,] = np.ravel(q)
+                    psit[i+1,] = np.ravel(psi)
+                    i+=1
+                u = np.concatenate([qt[i],psit[i]],0)
+                yobs[icyc,:,:3] = obsloc
+                yobs[icyc,:,3] = self.obs.add_noise(self.obs.h_operator(obsloc,u))
+            np.save(fq,qt)
+            np.save(fp,psit)
+            np.save(obsfile, yobs)
+        else:
+            logger.info("read truth")
+            qt = np.load(fq)
+            psit = np.load(fp)
+            if not os.path.isfile(obsfile):
+                logger.info("create obs")
+                for icyc in range(self.na):
+                    u = np.concatenate([qt[(icyc+1)*self.nt],psit[(icyc+1)*self.nt]],0)
+                    yobs[icyc,:,:3] = obsloc
+                    yobs[icyc,:,3] = self.obs.add_noise(self.obs.h_operator(obsloc,u))
+                np.save(obsfile, yobs)
+            else:
+                logger.info("read obs")
+                yobs = np.load(obsfile)
+            
         return qt, psit, yobs
 
     # initialize control 
@@ -120,6 +139,7 @@ class SO08_qg_func():
         q0 = np.zeros((self.nx,len(self.t0f)))
         psi0 = np.zeros((self.nx,len(self.t0f)))
         for j in range(len(self.t0f)):
+            logger.debug(self.t0f[j])
             inf = self.truedir+f"/q{self.t0f[j]:06d}.npy"
             q0[:,j] = np.ravel(np.load(inf))
             inf = self.truedir+f"/p{self.t0f[j]:06d}.npy"
@@ -155,20 +175,27 @@ class SO08_qg_func():
     # forecast
     def forecast(self,u):
         if self.ft == "ensemble":
-            q = u[:self.nx,:].reshape(self.ni,self.nj,-1)
-            psi = u[self.nx:,:].reshape(self.ni,self.nj,-1)
-            qf = np.zeros((self.a_window, self.nx, q.shape[2]))
-            psif = np.zeros((self.a_window, self.nx, psi.shape[2]))
+            nmem = u.shape[1]
+            qf = np.zeros((self.a_window, self.nx, nmem))
+            psif = np.zeros((self.a_window, self.nx, nmem))
+            for m in range(nmem):
+                q = u[:self.nx,m].reshape(self.ni,self.nj)
+                psi = u[self.nx:,m].reshape(self.ni,self.nj)
+                for l in range(self.a_window):
+                    for k in range(self.nt):
+                        q = self.step(q,psi)
+                    qf[l,:,m] = np.ravel(q)
+                    psif[l,:,m] = np.ravel(psi)
         else:
             q = u[:self.nx].reshape(self.ni,self.nj)
             psi = u[self.nx:].reshape(self.ni,self.nj)
             qf = np.zeros((self.a_window, self.nx))
             psif = np.zeros((self.a_window, self.nx))
-        for l in range(self.a_window):
-            for k in range(self.nt):
-                q = self.step(q,psi)
-            qf[l] = np.ravel(q)
-            psif[l] = np.ravel(psi)
+            for l in range(self.a_window):
+                for k in range(self.nt):
+                    q = self.step(q,psi)
+                qf[l] = np.ravel(q)
+                psif[l] = np.ravel(psi)
         
         if self.a_window > 1:
             uf = np.concatenate([qf,psif],1)
@@ -178,38 +205,91 @@ class SO08_qg_func():
             return uf
 
     # plot initial state
-    def plot_initial(self, psi, qt, psit, yobs):
+    def plot_state(self, t, q, psi, qt, psit, yobs):
         plt.rcParams["font.size"] = 16
-        fig, axs = plt.subplots(nrows=1,ncols=2,figsize=(12,5))
-        z = [psit.reshape(self.ni,self.nj),qt.reshape(self.ni,self.nj)]
+        logger.info("plot true and analysis states")
+        logger.debug(t)
+        logger.debug(q.shape)
+        logger.debug(psi.shape)
+        logger.debug(qt.shape)
+        logger.debug(psit.shape)
+        logger.debug(yobs.shape)
+        zt = [psit.reshape(self.ni,self.nj).transpose(),qt.reshape(self.ni,self.nj).transpose()]
+        if self.pt == "mlef":
+            z = [psi[:,0].reshape(self.ni,self.nj).transpose(),q[:,0].reshape(self.ni,self.nj).transpose()]
+        else:
+            z = [psi.mean(axis=1).reshape(self.ni,self.nj).transpose(),q.mean(axis=1).reshape(self.ni,self.nj).transpose()]
         zmax = [6.0e1,1.5e5]
         x = np.linspace(0, 1, self.ni)
         y = np.linspace(0, 1, self.nj)
         title = [r"stream function $\psi$", r"potential vorticity $q$"]
-        for i in range(len(axs)):
-            ax = axs[i]
-            c = ax.pcolormesh(x, y, z[i],
-                   cmap="RdYlBu_r", vmin=-zmax[i], vmax=zmax[i])
-            ax.set_title(title[i])
+        fname = ["p","q"]
+        for i in range(len(z)):
+            fig, axs = plt.subplots(nrows=1,ncols=2,figsize=(12,5))
+            ax = axs[0]
+            c = ax.pcolormesh(x, y, zt[i],
+               cmap="RdYlBu_r", vmin=-zmax[i], vmax=zmax[i])
+            ax.set_title("truth")
             ax.set_aspect("equal")
             fig.colorbar(c, ax=ax, shrink=0.8)
-        axs[0].scatter(yobs[:,0]/self.ni,yobs[:,1]/self.nj,marker='o',
-                    c=yobs[:,2],s=3.0,edgecolors='k',
+            if i==0:
+                ax.scatter(yobs[:,1]/self.ni,yobs[:,2]/self.nj,marker='o',
+                    c=yobs[:,2],s=7.0,edgecolors='k',linewidths=0.3,
                     cmap="RdYlBu_r", vmin=-zmax[0], vmax=zmax[0])
-        fig.suptitle(r"$t=$"+f"{self.t0true}")
-        fig.tight_layout()
-        fig.savefig("qg_initial_truth.png")
+            ax = axs[1]
+            c = ax.pcolormesh(x, y, z[i],
+               cmap="RdYlBu_r", vmin=-zmax[i], vmax=zmax[i])
+            ax.set_title("analysis")
+            ax.set_aspect("equal")
+            fig.colorbar(c, ax=ax, shrink=0.8)
+            if t >= 0:
+                fig.suptitle(r"$t=$"+f"{t*self.nt*self.dt}, "+title[i])
+            else:
+                fig.suptitle("initial, "+title[i])
+            fig.tight_layout()
+            if t >= 0:
+                fig.savefig(f"{fname[i]}{t:06d}.png")
+            else:
+                fig.savefig(f"{fname[i]}initial.png")
+            plt.close()
 
         fig, axs = plt.subplots(nrows=5,ncols=5,figsize=(15,15),
         subplot_kw={'xticks': [], 'yticks': []})
         for i in range(len(axs.flatten())):
             ax = axs.flatten()[i]
-            p = psi[:,i].reshape(self.ni,self.nj)
+            p = psi[:,i].reshape(self.ni,self.nj).transpose()
             c = ax.pcolormesh(x, y, p,
                    cmap="RdYlBu_r", vmin=-zmax[0], vmax=zmax[0])
-            ax.set_title(r"$t=$"+f"{self.t0f[i]}")
+            if self.pt == "mlef":
+                ax.set_title("mem "+f"{i}")
+            else:
+                ax.set_title("mem "+f"{i+1}")
             ax.set_aspect("equal")
         #fig.colorbar(c, shrink=0.8)
         #fig.suptitle(r"$t=$"+f"{self.t0true}")
         fig.tight_layout()
-        fig.savefig("qg_initial_ens.png")
+        if t>=0:
+            fig.savefig(f"pens{t:06d}.png")
+        else:
+            fig.savefig(f"pensinitial.png")
+        plt.close()
+        fig, axs = plt.subplots(nrows=5,ncols=5,figsize=(15,15),
+        subplot_kw={'xticks': [], 'yticks': []})
+        for i in range(len(axs.flatten())):
+            ax = axs.flatten()[i]
+            p = q[:,i].reshape(self.ni,self.nj).transpose()
+            c = ax.pcolormesh(x, y, p,
+                   cmap="RdYlBu_r", vmin=-zmax[1], vmax=zmax[1])
+            if self.pt == "mlef":
+                ax.set_title("mem "+f"{i}")
+            else:
+                ax.set_title("mem "+f"{i+1}")
+            ax.set_aspect("equal")
+        #fig.colorbar(c, shrink=0.8)
+        #fig.suptitle(r"$t=$"+f"{self.t0true}")
+        fig.tight_layout()
+        if t>=0:
+            fig.savefig(f"qens{t:06d}.png")
+        else:
+            fig.savefig(f"qensinitial.png")
+        plt.close()
